requirement.txt
langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
gradio==3.28.3
gradio_client==0.1.4
requests==2.29.0
gTTS==2.3.1